fix: Fix PHP error in buy route
When the quantity is provided as non-numeric string, the method cannot be called because it only takes an int.

+++ site/routes/buy.php
@@ -60,7 +60,7 @@ return [
 			$postalCode = get('postalCode');
 			$state      = get('state');
 			$street     = get('street');
-			$quantity   = Product::restrictQuantity(get('quantity', 1));
+			$quantity   = Product::restrictQuantity((int)get('quantity', 1));
 			$vatId      = get('vatId');
 
 			try {
